+++ dotnet/src/SmartAmbience.MvcSupport/Threading/SmartAmbienceSetupExtensionsForApplicationBuilder.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.AspNetCore.Builder;
+using System.Threading;
+
+namespace Microsoft.AspNetCore {
+
+  public static class SmartAmbienceSetupExtensionsForApplicationBuilder {
+
+    public static void UseAmbientFieldAdapterMiddleware(this IApplicationBuilder app) {
+      app.UseMiddleware<AmbientFieldAdapterMiddleware>();
+      AmbientField.ContextAdapter = AmbientFieldAdapterMiddleware.ContextAdapter;
+    }
+
+  }
+
+}