+++ dotnet/src/SmartAmbience/(internal)/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System.ComponentModel;
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("SmartAmbience.Tests")]