+++ vers/changelog.md
@@ -8,7 +8,9 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
+* adjusted some small changes to the AmbientField
+* added the 'AmbientFieldAdapterMiddleware' (**new Feature**) (+setup extension) to configure a ASP .net core Webapi environment for appropriate usage of AmbientFields
+* added the 'ExtractFlowedDataAttribute' to automatically apply incoming flowed data (from HTTP-headers) into 'AmbientFields' (**new Feature**)